+++ service/src/main/java/gov/nasa/pds/api/registry/controllers/MyProductsApiBareController.java
@@ -26,10 +26,10 @@ import gov.nasa.pds.api.registry.business.LidVidNotFoundException;
 import gov.nasa.pds.api.registry.business.LidVidUtils;
 import gov.nasa.pds.api.registry.business.ProductBusinessObject;
 import gov.nasa.pds.api.registry.business.RequestAndResponseContext;
+import gov.nasa.pds.api.registry.elasticsearch.ElasticSearchRegistryConnection;
 import gov.nasa.pds.api.registry.exceptions.ApplicationTypeException;
 import gov.nasa.pds.api.registry.exceptions.NothingFoundException;
 import gov.nasa.pds.api.registry.search.ElasticSearchHitIterator;
-import gov.nasa.pds.api.registry.search.ElasticSearchRegistryConnection;
 import gov.nasa.pds.api.registry.search.ElasticSearchRegistrySearchRequestBuilder;
 import gov.nasa.pds.api.registry.search.KVPQueryBuilder;
