+++ service/src/main/java/gov/nasa/pds/api/registry/elasticsearch/ElasticSearchConfig.java
@@ -1,4 +1,4 @@
-package gov.nasa.pds.api.registry.search;
+package gov.nasa.pds.api.registry.elasticsearch;
 
 import java.util.List;
 import org.slf4j.Logger;
@@ -11,6 +11,7 @@ import org.springframework.context.annotation.Configuration;
 import gov.nasa.pds.api.registry.SystemConstants;
 import gov.nasa.pds.api.registry.business.ProductBusinessObject;
 import gov.nasa.pds.api.registry.configuration.AWSSecretsAccess;
+import gov.nasa.pds.api.registry.search.ElasticSearchRegistrySearchRequestBuilder;
 
 /* Keep this eventhough not directly referenced
  *
